+++ packages/eas-cli/src/graphql/queries/BuildQuery.ts
@@ -8,16 +8,28 @@ import {
   BuildStatus,
   BuildsByIdQuery,
   BuildsByIdQueryVariables,
+  DistributionType,
   GetAllBuildsForAppQuery,
   GetAllBuildsForAppQueryVariables,
 } from '../generated';
 import { BuildFragmentNode } from '../types/Build';
 
-type Filters = {
-  platform?: AppPlatform;
-  status?: BuildStatus;
+type BuildsQuery = {
   offset?: number;
   limit?: number;
+  filter?: {
+    platform?: AppPlatform;
+    status?: BuildStatus;
+    distribution?: DistributionType;
+    channel?: string;
+    appVersion?: string;
+    appBuildVersion?: string;
+    sdkVersion?: string;
+    runtimeVersion?: string;
+    appIdentifier?: string;
+    buildProfile?: string;
+    gitCommitHash?: string;
+  };
 };
 
 export const BuildQuery = {
@@ -52,7 +64,7 @@ export const BuildQuery = {
 
   async allForAppAsync(
     appId: string,
-    { limit = 10, offset = 0, status, platform }: Filters
+    { limit = 10, offset = 0, filter }: BuildsQuery
   ): Promise<BuildFragment[]> {
     const data = await withErrorHandlingAsync(
       graphqlClient
@@ -63,13 +75,12 @@ export const BuildQuery = {
               $appId: String!
               $offset: Int!
               $limit: Int!
-              $status: BuildStatus
-              $platform: AppPlatform
+              $filter: BuildFilter
             ) {
               app {
                 byId(appId: $appId) {
                   id
-                  builds(offset: $offset, limit: $limit, status: $status, platform: $platform) {
+                  builds(offset: $offset, limit: $limit, filter: $filter) {
                     id
                     ...BuildFragment
                   }
@@ -78,7 +89,7 @@ export const BuildQuery = {
             }
             ${print(BuildFragmentNode)}
           `,
-          { appId, offset, limit, status, platform }
+          { appId, offset, limit, filter }
         )
         .toPromise()
     );